Remove unused search location field.
The presence of a second, but hidden, field causes iOS Safari to not doing
anything after hitting the "return" key on the on screen keyboard. I suspect
this is because normally that key would take you to the next field, but because
it's invisible, it does nothing.

On Android Chrome, it appears that invisible fields are skipped and hitting
"return" will submit the form.