Simplify z-index handling. (#1028)
In order to prevent an arms race of larger z-indexes, this change
removed z-indexes from elements that do not need them and reduces
the z-index values of elements that do need them. It also adds some
z-index: 0 values (with comments!) in order to explicitly construct a
stacking context to prevent descendant elements from hovering over
higher level sibling elements.

The remaining z-indexes are between 0 and 3.