Migrate home page components to TypeScript (#1180)
* Migrate HomePage components to TypeScript.

The one change that is not just adding type annotations was to change
the `openTypeform()` function to not take in its first argument, `e`,
since it's not actually used.

* Upgrade Typeform to get TypeScript definitions.

The version of Typeform we were using was very old, and the newer
versions com with TypeScript definitions as part of the package. This
upgrades the `@typeform/embed` dependency to the latest version and
makes minor changes to the code to match the newest APIs.