Allow Texting modal to be scrollable.
This is important for shorter screens (< ~700px), where the submit
button on the Texting modal may not be visible.

This also adds a somewhat unrelated (though kind of related) fix to the
Texting modal, where the Filters bar from the search results page was
above the Texting modal along the z dimension. Forcing the main page and
the modal to be in different stacking contexts fixes this issue and
overall makes the stacking and scrolling a bit easier to reason about.