Migrate DataService to TypeScript.
This takes a shortcut by using `any` in many of the type signatures.
There's not much we can really do about this, since most of these
functions are generic and either call `JSON.stringify()` on an arguments
or `resp.json()` on responses. For better type safety, we probably want
the wrapper functions that call these ones to have stricter type
annotations.