Mmigrate APIHooks and DataService to TypeScript (#1214)
* Migrate DataService to TypeScript.

This takes a shortcut by using `any` in many of the type signatures.
There's not much we can really do about this, since most of these
functions are generic and either call `JSON.stringify()` on an arguments
or `resp.json()` on responses. For better type safety, we probably want
the wrapper functions that call these ones to have stricter type
annotations.

* Migrate APIHooks to TypeScript.

These use type assertions on the API responses, since we do not do any
runtime validation of data coming from the API. This also fixes a bug in
the ServiceDiscoveryResults type signatures, where it was expecting a
type of CategoryRefinement, but in reality we only had the data from a
Category (no Algolia-specific data in it).