Convert Edit page files to .tsx with weak types. (#1230)
This converts all of the Edit page JSX files to TSX, but without adding
the necessary type annotations everywhere. Instead, this opts out of
lint rules and even downgrades the tsconfig settings in order to make
minimal code canges.

Note that the tsconfig.json changes to disable `noImplicitAny` affect
the whole codebase; there's no way to change the tsconfig settings for a
subset of files without creating a separate tsconfig project. This will
only be temporarily enabled until we add stronger types to the Edit page
code.