Migrate Edit Page to TypeScript Part 3 (#1236)
* Edit Page: Add type for frontend internal version of Schedule.

The frontend app has a slightly different representation of Schedules
and ScheduleDays due to different needs in the UI. The code already
operates this way, but this adds a explicit types for these internal
structures.

* Edit Page: Add types for the transformed versions of Organization and Service.

The Edit Page has to maintain state about what changes were made, and
therefore what changes need to be synced to the API, so the models for
Organization and Service are not 100% identical to the versions that
come from the API.

Besides type annotations and assertions, there is one other larger
change to the code, which is that we default to the empty string before
passing a value to the `defaultValue` attribute on the form `<input>`
elements. We were failing to type check here because these values can
sometimes be `null`, which is not allowed by the builtin type
definitions for HTML attributes. The online documentation doesn't really
say what happens when you pass in a null, and from messing around in the
Chrome debug console, it looks like it basically treats it as if it were
the empty string. We now explicitly coerce `null` and `undefined` to the
empty string to match that behavior without tripping up undefined
behavior.