Add types to EditSidebar.
This required one non-type change to the OrganizationEditPage, which is
that in the main `render()` function, the `<Loader>` component is only
shown if `!resource` rather than when `!resource && !newResource`.

The reason for this change is that it allows TypeScript to infer that
`resource` is not undefined in the else block, which we need in order to
pass `resource` to the other components in that half of the `render()`
function. This change should be safe because the only time we ever set
`newResource` to `true`, at the same time we also set `resource` to a
value. This in turn means that, by the contrapositive, when `resource`
is undefined, then `newResource` must be false.