Edit Page: Add type annotations to `applyChanges()`.
Upon closer inspection of this function, it is less general than it
appears, since it makes some assumptions about the the inputs that are
very specific to how Services are represented on the Edit Page. As such,
it was given a type signature that includes both of the internal Service
types.

This makes one functional change to `applyChanges()`, which was to
remove the case where we operate on an item in `baseItems` that is not
present in `changesById`. This is not actually possible to hit today,
since we always prepopulate `changesById` with every service that was
originally on the Organization that comes from the API, and we rely on
this assumption later to ensure that certain fields always exist on the
return value of `applyChanges()` (e.g. the `scheduleObj` field, which is
not present on `InternalOrganizationService`). We now throw an exception
in this case to signal to TypeScript that this is not possible to hit.