Migrate Edit Page to TypeScript Part 4 (#1237)
* Edit Page: Split InternalService into two separate types.

It appears that the shape of an InternalService that is at the top level
of the React Component's state is slightly different from an
InternalService that is nested under the resources
(InternalOrganization) object.

This splits InternalService into InternalTopLevelService and
InternalOrganizationService.

* Add types to EditSidebar.

This required one non-type change to the OrganizationEditPage, which is
that in the main `render()` function, the `<Loader>` component is only
shown if `!resource` rather than when `!resource && !newResource`.

The reason for this change is that it allows TypeScript to infer that
`resource` is not undefined in the else block, which we need in order to
pass `resource` to the other components in that half of the `render()`
function. This change should be safe because the only time we ever set
`newResource` to `true`, at the same time we also set `resource` to a
value. This in turn means that, by the contrapositive, when `resource`
is undefined, then `newResource` must be false.

* Edit Page: Add type annotations to `applyChanges()`.

Upon closer inspection of this function, it is less general than it
appears, since it makes some assumptions about the the inputs that are
very specific to how Services are represented on the Edit Page. As such,
it was given a type signature that includes both of the internal Service
types.

This makes one functional change to `applyChanges()`, which was to
remove the case where we operate on an item in `baseItems` that is not
present in `changesById`. This is not actually possible to hit today,
since we always prepopulate `changesById` with every service that was
originally on the Organization that comes from the API, and we rely on
this assumption later to ensure that certain fields always exist on the
return value of `applyChanges()` (e.g. the `scheduleObj` field, which is
not present on `InternalOrganizationService`). We now throw an exception
in this case to signal to TypeScript that this is not possible to hit.

* Add types to EditServices component.

Note that this leaves `resourceAddresses` as a `Record<any, any>[]`,
since there is some more work required to determine the correct type for
that state variable.