Edit Page: Add type annotations to ProvidedService. (#1247)
This adds TypeScript type annotations to the ProvidedService component.

In order for the types to validate, this required modifying the
definition of InternalFlattenedService to make the `schedule` property
required. This appears to be true in practice, where either the
`schedule` property is defined by the original API response, or it gets
set by the frontend code when creating a new Service, so there shouldn't
be a case where it is undefined.

This adds an additional runtime check to the `applyChanges()` method to
assert that the final, flattened service indeed has a `schedule`
property.