Edit Page: Add types to EditAddress components.
This adds strict types to anything related to addresses, including the
EditAddress component and the `addresses` field in the
OrganizationEditPage's state variable.

It introduces a new InternalAddress type, which extends the API Address
type but adds a couple of properties for keeping track of modified
Addresses or Addresses that are intended to be removed.