Edit Page: Add type annotations to EditNotes. (#1262)
This is a mostly straightforward change to add strict type annotations
to anything related to editing notes. There were some surprising things
discovered from adding strict types, mostly involving the direct
mutation of objects that should not be mutated in React. We leave this
behavior in for now, but with comments describing what should eventually
be fixed.