Edit Page: Add type annotations to EditPhone and EditCollection.
This adds type annotations to the EditPhone component as well as the
editCollectionHOC function. Note that there are some things in
editCollectionHOC that cannot be correctly typed, since it appears to be
incorrectly calling `setState()`, passing in the array of collection
items rather than an object with a `collections` field. We leave the
existing code alone, since it may be a little bit of work to untangle
the technical debt, but we at least add stronger type annotations to
everything around it.