Restore Service.documents, but delete postDocuments().
We recently removed Service.documents ("patient handouts") from the Edit
Page, since it was no longer being read from the non-edit pages of the
site. However, we still need knowledge of the `documents` field on the
Edit Page, since we need to delete it off the Service object before
POSTing it to the API, since the POST API endpoint only allows you to
post columns of the Service table, and the documents are actually in a
separate DB table.

We restore the Service.documents field as an `unknown[]`, which allows
us to retain the knowledge of the field so that we know to delete it,
but without providing any detailed information about it.

Finally, we remove the postDocuments() function from the Edit Page,
since it is unreachable now that the form fields have been removed from
the Edit Page.