ProvidedService: Split up expression to add explicit type assertion.
TypeScript's type system is unable to provide exact types for
`Object.entries()`, since TypeScript only documents the minimum set of
properties that are expected on an object, but in practice an object may
have more properties. In this particular case, we are pretty sure that
we are not adding any more properties, so we just assert that the keys
and values are what we expect.