Re-enable tsconfig `noImplicitAny` (#1283)
* WeGlot: Add explicit type annotations to `handleChangeClientState()`.

* EditPhones: Add explicit type assertion for the `field` property.

* ProvidedService: Split up expression to add explicit type assertion.

TypeScript's type system is unable to provide exact types for
`Object.entries()`, since TypeScript only documents the minimum set of
properties that are expected on an object, but in practice an object may
have more properties. In this particular case, we are pretty sure that
we are not adding any more properties, so we just assert that the keys
and values are what we expect.

* OrganizationEditPage: Add explicit type annotations.

* Re-enable tsconfig's noImplicitAny.

This was originally disabled during the migration of the Edit Page to
TypeScript, since it would have been laborious to not violate this rule.
Now that the Edit Page has been fully migrated, we can re-enable this
check.