Remove unused google.js.
This does not appear to be used at all, and because it's not even
imported by any of our source files, it doesn't even appear in the final
JS bundle. This looks like it used to be imported for a global side
effect of including the Google Maps JavaScript file at the bottom of the
page, but we currently use React-specific libraries for embedding Google
Maps on our site today.