Style Sign-Up and Log-In pages.
This styles the sign-up and log-in pages to mostly match their designs.
There were no design mockups for mobile, so this includes a best-effort
attempt at defining styles for mobile.

As part of this change, this also moves the `Partners` component out of
the `HomePage` directory and to the common components directory, since
that component is now intended to be reused on these new pages. Also,
this removes a bit of margin from the component's definition, since the
sign-up and log-in pages shouldn't have that much margin, and it has
been manually added back to the `HomePage` in the form of an ad hoc
spacer div.