NavigatorDashboard: Fix infinite rerender bug.
Since we haven't filled out the `PARENT_ELIGIBILITIES` variable yet, the
current logic keeps trying to fetch the child eligibilities for them,
setting the `eligibilities` state variable to the empty array over and
over again. If we actually set values for the `PARENT_ELIGIBILITIES`,
then this wouldn't be a problem, but for now, guard the actual calling
of `setEligibilities` to avoid infinite rerendering.