[AuthService] Rename Auth0-specific ID field.
We have two different notions of user IDs: one from Auth0, which is a
string and is contained in the JWT, and one from our own API, which is
an integer. This renames the `id` field within the auth state to
`externalId`, to make it clearer that these are from the identity
provider.