Hook up bookmarks APIs. (#1366)
This replaces all the placeholder data related to bookmarks with actual
API calls to the newly added Bookmarks and Folders APIs. There are a few
mismatches between the existing frontend code and the backend APIs that
we actually ended up implementing. For expediency, we mostly just work
around gaps or omissions from the backend APIs, but ideally these should
be addressed in the future.

Two recurring issues are that the Bookmarks API does not have a name
field, and it does not return the `updated_at` field. The former doesn't
even exist in the database table definition, but the latter is just an
omission. Some of the changes in this commit work around the lack of
these fields, including making an API request to the bookmarked service
or resource to grab its name.

There are some awkward places where we make a large number of API calls
to fetch information that isn't easily accessible, such as getting the
largest bookmark `order` number, which we can only do by querying all
bookmarks owned by the user.

Finally, there is some awkwardness due to the fact that the user's ID is
not easily accessible in the frontend code. The JWT from Auth0 only
contains the "external ID" coming from Auth0, and our authentication
code is too complicated to easily slip in our API's user IDs into the
same React context, mostly due to needing to correctly handle the case
where a user is signing up or simply logging in. We opt for the safer,
though awkward strategy of always querying the API for the user's ID
just before making any API call that actually requires the user ID. This
ensures that we will always have access to the user ID, at the cost of
requiring an additional API request for every actual API request we
really want to make.