Fix outdated names in refinement mappings.
The refinement mappings are a hardcoded list of eligibility and
category names used for the left sidebar filters on the search results
page (but not the tile results page). This list has been neglected, and
various eligibilities and categories have been renamed or deleted from
the database. This commit updates the names based on feedback and advice
from the Content team, even though some of these even predate the
Content team.