Hook up saved search APIs.
This hooks up the placeholder locations in the Navigator UI with the
saved search APIs.

When a user is logged in, we add a "Save Results" button to the search
results page, and pressing that button will fire an API request to save
the current search filters and query to the database. This involved
pulling out the pieces of the search state object that Algolia's
InstantSearch API manages into the format our API accepts.

On the main dashboard page, we perform the reverse operation, querying
our API for saved searches, and then reconstituting a simulacrum of the
serach state object, solely for the purposes of generating the URL query
parameters directly from that object, which is how the search results
page normally works.