Add login/profile widget to nav bar.
This allows the login page to be discoverable, since previously, you had
to manually type in the address of the login page into the address bar.
This is gated as a new whitelabel property named `loginEnabled`, which
is only being enabled on the SFSG site and the legacy default askdarcel
site.