Refresh BookmarksMenu when it is opened. (#1378)
Without adding this logic, the BookmarksMenu will not reflect any new
bookmarks that are added unless you refresh the browser. The reason is
because the menu is always mounted, as it lives at the `App` layer,
though hidden off screen. This means that the `useEffect()` call only
happens once, when the app is first loaded.

By capturing the `isOpen` variable in the `useEffect()`'s dependency
list, we can retrigger the `useEffect()` call whenever the menu is
opened, causing it to be updated with the latest state of the user's
bookmarks.