Enable Breaking News banner for heat alert.
There is currently a heat advisory, and we want to link to some timely
notices from the City of San Francisco in the SFSG whitelabel site. This
enables the Breaking News Banner and updates the previous text, which
was for a cold weather alert, to now mention the hot weather information
and an up-to-date link to the City's website.