Boost service discovery (tile) results based on boosted category. (#1392)
In order to surface some specific results on the tile results page, we
make use of an optional filter on a newly added Algolia index property.

On the backend, we've created a new column, `boosted_category`, which we
can optionally set on each Service. This allows us perform an optional
filter only on the tile results pages, and not the general search page,
where services that match the optional filter are boosted in ranking.

For now, there can only be one `boosted_category` per service, but this
is sufficient for the current use case, where only two services need to
be boosted, and they each only apply to one category.