Fix Material Icons by migrating to Material Symbols.
Something appears to have changed about the Material Icons icon font
that we're loading from Google Fonts, which is making them render
strangely. Most severely is that on iOS, the width of the element
containing the icons is 1,000,000 pixels wide, which not only stretches
the page but also blocks click events from going through, even though
the element itself is transparent. On desktop and Android, we do not see
the extremely wide icons, but it appears that the icons are noticeably
smaller than they used to be.

Several years ago, Google released Material Symbols as a successor to
Material Icons. Switching our app to Material Symbols appears to get rid
of the issues described above.

Note that there are slight differences in the icons after this change.
Most of the differences are due to icons changing from being outlined
(no fill) to filled with a solid color. Material Icons had separate icon
names for any icons with different fill vs. outlined variants, but
Material Symbols controls this with the font URL query and the
`font-variant-settings` CSS property. This means that when migrating
from Material Icons, any icons with names like `foo_outline` must be
renamed to `foo`. We don't bother with loading both the outlined and
filled variations of the font, so we've instead migrated several icons
to the filled variant.