Fix local development webpack server. (#1438)
It's likely that the recent changes to the webpack build to add the hash
of the bundle.js file broke something about the local dev server. This
fixes the configuration so that assets can be served out of both the
root URL and `/dist`. This also required upgrading the version of
webpack-dev-server to be able to use the new options.